Moved `View` to abstract_ui Package
