+++ app/src/main/java/xterminators/spellingbee/abstract_ui/View.java
@@ -1,0 +1,3 @@
+package xterminators.spellingbee.abstract_ui;
+
+public abstract class View {}
