--- app/src/main/java/xterminators/spellingbee/view/View.java
@@ -1,3 +1,0 @@
-package xterminators.spellingbee.view;
-
-public abstract class View {}
